Data Validation set for cells
The following restrictions are set for the different columns:
For all the dates, a range between 2000-01-01 to 2999-12-31 is set and only valid date format will be accepted;
Different range and format(integers only) restrictions are also set for period(1-5), week(1-20), Nr(1-100), Nr of exercise(1-100) and Weight(0-100).

When the input value doesn’t match the data validation restrictions defined for the cell, an alter would occur to remind the user to cancel or retry.