Add protections to particular cells
The following changes happened to different sheets after protection operation:

Course: Only three cells are editable(namely A2, B2 and C2), the rest cells are locked.
Lectures: Only Row 2 to 1000 of Column A to E are editable while Columns from F and beyond (unused columns)are locked.
Assignments:Only Row 2 to 1000 of Column A to I are editable while Column from J and beyond (unused columns)are locked.
OldExam: Column B,I,J and beyond O(not including O) as well as all the lines after 1000 in other columns(which needed to be filled with sth) are locked. The other columns within 1000 lines are editable.
Exercises: Only A2:A1000, B2:B1000, C2:C1000, D2:D1000 and E2:E1000 are editable, the rest cells are locked.
StudyMaterials: Only Row 2 to 1000 of Column A to F are editable, the others are locked.
Exams: Only Row 2 to 10 of Column A to D are editable, the others are locked.

FYI: in case of change the protection areas, the following link gives instructions on how to protect selected cells in Excel
https://support.office.com/en-us/article/lock-cells-to-protect-them-cb7835f6-9c37-4161-bb53-d1c410acaf21